refactor(tests): split PatchCoverageBoostTest into properly-named files
The 'boost' bucket file was sloppy naming — every WSRouter
validation / accessor test in it duplicated tests already present
in tests/Unit/WSRouterServerRegistryTest.php and
tests/Unit/WSRouterV040Test.php (and equally for Cache, Counter,
RedisPubSub, RedisStreams — checked one-by-one). Re-emitting the
same test under a different name DOES nothing for codecov/patch
because each line was already counted as hit.

Drops:
- tests/Unit/PatchCoverageBoostTest.php  (71 tests, ~95% duplicates)
- tests/Unit/WSRouterValidationTest.php  (intermediate split — every
  test in it was already covered by WSRouterServerRegistryTest +
  WSRouterV040Test)

Keeps the GENUINELY-new tests under feature-scoped names:
- tests/Unit/WSRouterClientDispatchTest.php  (5 tests) — sendToClient,
  broadcast Table-backend throw, pushWithBackpressure dead-fd via
  isEstablished-stubbed WS\Server subclass, Room pre-init guard
- tests/Unit/WS/CapacityExceptionTest.php    (3 tests) — type
  hierarchy, message round-trip, catchable-as-StoreException
- tests/Unit/Store/DefaultBackendConfigTest.php (16 tests) — Store +
  Counter ::defaultBackend memcached env + array configs, redis array
  configs (prefer string/enum/invalid fallback), tiered array configs,
  DriverPreference coerce semantics (no prior DriverPreference test)
- tests/Unit/CacheFileTierAndStampedeTest.php (2 tests) — gcFiles
  eviction loop with externally-set maxFiles cap (was only
  smoke-tested), getOrCompute stampede stall (lock taken externally
  to drive the 200ms poll-then-fall-through path)

Appended to tests/Unit/AppPubSubAliasesTest.php:
- testPublishThrowsOnTableBackend
- testPublishReliableThrowsOnTableBackend
(prior tests only verified the method exists, not the behaviour)

PHPStan level-10 clean, 35 tests / 59 assertions green.
